User login
This is really a brand new repo. The backend is still in Java but this new frontend includes Jest (RTL), DaisyUI (Tailwind), TypeScript, React 18, and NextJS. This is going to make it much easier to develop on and get a much faster feedback loop. The next step is to hoist the login to the design system and then work on the character sheet. That also needs to go into the design system. More work to come.